修复autherror样式
