+++ app/src/main/java/org/lntu/online/ui/base/DrawerLayoutActivity.java
@@ -8,17 +8,7 @@ import android.view.WindowManager;
 
 import org.lntu.online.ui.widget.ThemeUtils;
 
-public abstract class DrawerLayoutActivity extends BaseActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT && Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-            getWindow().addFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
-        } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            getWindow().getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_STABLE | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN);
-        }
-    }
+public abstract class DrawerLayoutActivity extends FullLayoutActivity {
 
     public void adaptStatusBar(View statusBar) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
