+++ app/src/main/res/values/styles.xml
@@ -17,7 +17,12 @@
 
     <style name="AppTheme.Launch" />
 
-    <style name="AppTheme.Translucent" parent="@android:style/Theme.Translucent.NoTitleBar" />
+    <style name="AppTheme.Translucent">
+        <item name="android:windowBackground">@android:color/transparent</item>
+        <item name="android:colorBackgroundCacheHint">@null</item>
+        <item name="android:windowIsTranslucent">true</item>
+        <item name="android:windowAnimationStyle">@android:style/Animation</item>
+    </style>
 
     <!-- App Theme Dialog -->
     <style name="AppDialog" parent="Theme.AppCompat.Light.Dialog">
